Separated into Example and Testing activity, and added new launcher icon
