+++ app/src/main/res/layout/activity_config_change_test.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              xmlns:tools="http://schemas.android.com/tools"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:animateLayoutChanges="true"
+              android:gravity="center_horizontal"
+              android:orientation="vertical"
+              android:padding="@dimen/activity_vertical_margin"
+              tools:context=".DemoActivity">
+
+    <Button
+        android:id="@+id/causeActivityPauseButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/cause_activity_pause"/>
+
+    <Button
+        android:id="@+id/causeLanguageChangeButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/cause_language_change"/>
+
+    <Button
+        android:id="@+id/causeOrientationChangeButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/cause_orientation_change"/>
+
+    <Button
+        android:id="@+id/loadDataButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/load_data"/>
+
+    <EditText
+        android:id="@+id/requestDurationEditText"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:gravity="center"
+        android:hint="@string/hint_request_duration"
+        android:inputType="numberSigned"/>
+
+    <ProgressBar
+        android:id="@+id/progressBar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:indeterminate="true"
+        android:visibility="gone"/>
+
+    <TextView
+        android:id="@+id/resultTextView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:freezesText="true"
+        android:gravity="bottom"
+        android:visibility="gone"/>
+
+</LinearLayout>
